wltc-data: FIX velocity-profile mistakenly rounded to integers! * preproc-utils: Minor enhancements on pandas and scripting. Experiment: Calculate in float64 (instead of 32).

+++ wltc/experiment.py
@@ -147,7 +147,7 @@ class Experiment(object):
 
         from .instances import wltc_data
 
-        self.dtype = np.float32
+        self.dtype = np.float64
         self.model = model
         self.wltc = wltc_data()
